website with almost full implementaiton
everything should be good, just not sure if the sending the LLAT and IP works and how we want me ot acc send it. currently im just sending them all when the record button is sent, including LLAT and IP.

i didnt add delete for SCriptID,LLAT, and IP yet since im not sure how we actually want that done.